HTTP replaced by HTTPClient, code refactored
