+++ TODO
@@ -7,15 +7,15 @@
 Todos:
     New:
         Current:
-            --        
+            ☐ Update tests for HttpClient
+
         ☐ While editing Todo use next animation - zoom in current Todo and shadowing backgroung        
         ☐ Write unit tests for events (emit) (look here - https://angular.io/guide/testing)
             ☐ Write tests for TodoListItemComponent
             ☐ Write tests for TodoListFooterComponent
             ☐ Write tests for TodoListComponent
             ☐ Write tests for TodoListHeaderComponent 
-        ☐ Rewrite unit tests for app.component.ts (exlude NO_ERRORS_SCHEMA)
-        ☐ Rewrite HTTP request to HTTPCommon
+        ☐ Rewrite unit tests for app.component.ts (exlude NO_ERRORS_SCHEMA)        
         ☐ Fix all TODO in all files
         ☐ After finishing the tutor, improve the app using (https://github.com/tastejs/todomvc-app-template)
         ☐ Find the way to start json server on gh-pages (look here https://my-json-server.typicode.com/) -> No 'Access-Control-Allow-Origin' header is present on the requested resource
@@ -26,6 +26,7 @@ Todos:
         ✔ Update unit tests for part-3
         ✔ Update unit tests for part-4
         ✔ Run npm run json-server and ng serve in the same terminal (look here https://github.com/kimmobrunfeldt/concurrently)
+        ✔ Rewrite HTTP request to HTTPClient
 
     Canceled:
         ✘ Combine two storing data method (HTTP and InMemory) -> Not necessarily anymore
