+++ db.json
@@ -16,19 +16,19 @@
       "id": 3
     },
     {
-      "title": "Finish",
-      "complete": false,
-      "id": 7
-    },
-    {
       "title": "Do all todos",
       "complete": false,
       "id": 8
     },
     {
       "title": "Pass upwork verification",
-      "complete": false,
+      "complete": true,
       "id": 9
+    },
+    {
+      "title": "Finish the project",
+      "complete": false,
+      "id": 10
     }
   ]
 }
