+++ src/app/todos.resolver.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core';
 import { Resolve, ActivatedRouteSnapshot, RouterStateSnapshot } from '@angular/router';
 import { Observable } from 'rxjs/Observable';
-import { TodoService } from './todo.service';
+import { TodoService } from './services/todo.service';
 import { ToDo } from './to-do';
 
 @Injectable()
