+++ src/app/todos/todos.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import { TodoService } from './../todo.service';
+import { TodoService } from './../services/todo.service';
 import { ToDo } from './../to-do';
 import { ActivatedRoute } from '@angular/router';
